+++ app/panel/directives/projectInlineFile/projectInlineFile.js
@@ -8,13 +8,15 @@
       link: function ($scope) {
         $scope.fileRoot = appSettings.fileRoot;
         $scope.removeFile = function () {
-          filesService.removeFile($stateParams.projectId, $scope.file.id)
-          .success(function () {
-            toaster.pop('success', 'File removed!', 'You have successfully removed the file.');
-          })
-          .error(function () {
-            toaster.pop('error', 'An error occured.', 'Unfortunately we couldn\'t remove the file. Please, try again');
-          });
+          if(confirm('Are you sure you want to remove the file?')) {
+            filesService.removeFile($stateParams.projectId, $scope.file.id)
+            .success(function () {
+              toaster.pop('success', 'File removed!', 'You have successfully removed the file.');
+            })
+            .error(function () {
+              toaster.pop('error', 'An error occured.', 'Unfortunately we couldn\'t remove the file. Please, try again');
+            });
+          }
         };
       }
     };
